+++ test/mui_suggestors/components/chip/label_to_mui_chip_migrator_test.dart
@@ -18,7 +18,7 @@ import 'package:test/test.dart';
 import '../../../resolved_file_context.dart';
 import '../../../util.dart';
 import '../shared.dart';
-import '../shared_props_migrators/color_props_test.dart';
+import '../shared_props_migrators/color_props.dart';
 import 'shared.dart';
 
 void main() {
